Merge branch 'master' of https://github.com/AhmedSalemElhady/Scrabble-Game.ai into IMP-Score-Evaluation/heuristics-calculation

+++ src/SharedClasses/Move.hpp
@@ -35,7 +35,7 @@ class Move
 
 public:
     std::string word; // Extra Notes : The Returned word will be lowered Case and a Capitalized Letter in case of a BLANK USED.
-    int moveScore;    // type maybe changed...
+    double moveScore;    // type maybe changed...
     Position startPosition;
     bool horizontal; // true -> horiz false -> vertical.
     double testHScore;
